requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
dirtyjson==1.0.7
idna==3.3
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.11
